git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@2281 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/thread-private.h
@@ -108,7 +108,7 @@ static inline size_t GetOpenMPMaximumThreads(void)
 #endif
 }
 
-static inline ssize_t GetOpenMPThreadId(void)
+static inline int GetOpenMPThreadId(void)
 {
 #if defined(MAGICKCORE_OPENMP_SUPPORT) && (_OPENMP >= 200203)
   return(omp_get_thread_num());
